Be less optimistic about using something else than gearman
Gearman has a pretty rich interface already, compared to a lot of theother existing job queues. Even while we don't use all of what Gearmanoffers, it's more than many even provide out of the box, and these jobqueues would therefore need to be shimmed to match.There are also a bunch of generic message queues in use as job queueswhich would therefore need even more shimming to make compatible.So it's not such a simple task to write a different backend. I'll stilltry to make it possible, but I won't lie about the effort likely needed.